Split AlertingPage (thing with tabs) out into a separate file
